+++ changelog.md
@@ -1,6 +1,11 @@
 vX.X.X
 ------
 
+- Javascript is loaded via the `WebView.executeScript()` method, closes #183.
+
+v0.8.4
+------
+
 - Updated ANSI escape sequence support .gif on home page, closes #180.
 - Improved the way that WebView is found to be "ready", closes #182.